requirements/base.txt
django>=1.6,<1.8
django-cms>=3.0.1
django-shop>=0.2.0
django-filer>=0.9.5
django-mptt>=0.6.0
django-parler>=1.4
measurement>=1.7.2
django-currencies>=0.3.2
